Fix database config for non-devcontainer environments
Back in this [commit](https://github.com/maybe-finance/maybe/commit/5f50ea3f02bfc1797816e3c679065c611fc61347#diff-5a674c769541a71f2471a45c0e9dde911b4455344e3131bddc5a363701ba6325R22-R23) I added a default postgres user and password.

Often this is not the user and password that is used in non-devcontainer environments.

This commit uses environment variables that are set in the devcontainer to set the user and password in the database.yml file.

If the environment variables are not set, it falls back to the default user and password on the users system.